♻️  router session_persistence.py 추출 — extra mutation + load helpers (P0-P step 6)
변경 이유
- 14 helper (``_persist_*`` 7 종 + ``_load_session_by_id`` /
  ``_most_recent_session`` + ``_is_terminal`` + ``_work_report_to_dict``
  + ``_record_persistence_failure`` + ``_proposal_to_dict/from_dict``)
  가 모두 ``session.extra`` 의 write surface 인데 _legacy 안 4 곳에
  흩어져 있었다. 모든 gate / hook 이 의존하므로 하나의 mid-level
  module 로 통합.

주요 변경 사항
- 새 파일 ``engineering_channel_router/session_persistence.py`` (약 530 줄):
  - lifecycle bookkeeping (role_selection / lifecycle_mode /
    coding_session_context / extra_keys / thread_id / terminal)
  - coding flow (proposal / job / serialization round-trip)
  - reporting (work_report dict serialization)
  - session lookup (by_id / most_recent)
- ``_legacy.py`` 의 해당 14 함수 (409 줄 합계) 제거 → 단일 통합
  re-export 만 남김.
- ``__init__.py`` facade 가 ``.session_persistence`` 에서 외부에서
  patch 하는 4 symbol (``_load_session_by_id`` / ``_most_recent_session``
  / ``_persist_lifecycle_mode`` / ``_persist_thread_id``) 직접 import.

비고
- 4672 PASS + 1 skipped 회귀 유지.
- 향후 session.extra 키 추가 / WorkReport schema 확장 시 본 모듈
  한 곳에서.
